+++ container_registry.go
@@ -96,7 +96,7 @@ type RegistryRepository struct {
 	CreatedAt              *time.Time               `json:"created_at"`
 	CleanupPolicyStartedAt *time.Time               `json:"cleanup_policy_started_at"`
 	Status                 *ContainerRegistryStatus `json:"status"`
-	TagsCount              int                      `json:"tags_count"`
+	TagsCount              int64                    `json:"tags_count"`
 	Tags                   []*RegistryRepositoryTag `json:"tags"`
 }
